+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **[Feature]** Expose custom registries for each target.
 - **[Feature]** Add `dist.tscOptions` for `lib` target to override options for
   distribution builds.
+- **[Feature]** Native ESM tests with mocha.
 - **[Fix]** Disable deprecated TsLint rules from the default config
 - **[Fix]** Remove use of experimental `fs/promises` module.
 - **[Internal]** Fix continuous deployment script (stop confusing PRs to master